Finish documentation (without exceptions) and rearrange files by packages
