--- src/main/kotlin/ru/hse/ezh/Executor.kt
@@ -1,10 +1,0 @@
-package ru.hse.ezh
-
-import java.io.InputStream
-
-object Executor {
-
-    fun execute(commands: List<Operation>, globalEnv: Environment): Triple<Int, InputStream, InputStream> =
-        TODO("Not yet implemented")
-
-}
