--- src/main/kotlin/ru/hse/ezh/Lexer.kt
@@ -1,9 +1,0 @@
-package ru.hse.ezh
-
-object Lexer {
-
-    fun lex(input: Sequence<Char>): List<Token> = TODO("Not yet implemented")
-
-    fun postprocess(tokens: List<Token>, globalEnv: Environment): List<Token> = TODO("Not yet implemented")
-
-}
