+++ src/main/kotlin/ru/hse/ezh/exceptions/CommandNotFoundException.kt
@@ -1,0 +1,3 @@
+package ru.hse.ezh.exceptions
+
+class CommandNotFoundException(message: String) : Exception(message)
