+++ src/main/kotlin/ru/hse/ezh/exceptions/ViewException.kt
@@ -1,0 +1,3 @@
+package ru.hse.ezh.exceptions
+
+class ViewException(message: String, cause: Throwable? = null) : Exception(message, cause)
