+++ src/main/kotlin/ru/hse/ezh/execution/commands/WcCommand.kt
@@ -1,7 +1,7 @@
-package ru.hse.ezh.commands
+package ru.hse.ezh.execution.commands
 
-import ru.hse.ezh.Command
 import ru.hse.ezh.Environment
+import ru.hse.ezh.execution.Command
 import java.io.InputStream
 import java.io.OutputStream
